Refactor diagnostic logic into a diagnostics service
